requirements.txt
flask==2.3.3
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
unidecode==1.3.6
python-dotenv==1.0.0
schedule==1.2.0